Changed Projet Structure to refactor code
